created tokens file on middelwares folder, standarized code
